move calculations services to shared
