+++ src/app/shared/shared-analysis/calculations/analysis-calculations-helper.service.ts
