+++ src/app/shared/shared-analysis/calculations/analysis-calculations.service.ts
